fix(test): corrige la estructura de recursos para las pruebas
